requirements.txt
Pillow
pytest
aiohttp
aiofiles
